requirements.txt
click==8.1.7
dataclasses_json==0.6.7
Flask==3.1.0
Flask_Bootstrap==3.3.7.1
Flask_Moment==1.0.6
flask_wtf==1.2.2
gpiozero==2.0.1
mfrc522==0.0.7
pydantic==2.10.6
pytest==8.3.4
python-dotenv==1.0.1
python_vlc==3.0.21203
Requests==2.32.3
RPi.GPIO==0.7.1
setuptools==70.0.0
SQLAlchemy==2.0.36
WTForms==3.2.1
pre-commit==4.1.0
